+++ src/ballistica/shared/ballistica.cc
@@ -39,7 +39,7 @@ auto main(int argc, char** argv) -> int {
 namespace ballistica {
 
 // These are set automatically via script; don't modify them here.
-const int kEngineBuildNumber = 22436;
+const int kEngineBuildNumber = 22442;
 const char* kEngineVersion = "1.7.44";
 const int kEngineApiVersion = 9;